ecommerce_project/ecommerce_project/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
defusedxml==0.7.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-xml==2.0.0
idna==3.10
oauthlib==3.3.1
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.5.0
